Compiler warnings fixes
Chopping a 64 bit int into a 32 bit.

+++ src/framework/text/stringtools.cpp
@@ -74,7 +74,7 @@ LIBTECH_API char** string_split(const char* source, const char* delim, int* resu
 
     }
 
-    *result_count = splitParts.Count();
+    *result_count = (int)splitParts.Count();
     char** res = splitParts.GetListData();
 
     return res;
